extend pfset
