+++ test/python/new_features/pfset_test/ComputationalGrid.yaml
@@ -1,0 +1,11 @@
+ComputationalGrid:
+  DX: 8.88888888888889
+  DY: 10.666666666666666
+  DZ: 1.0
+  Lower:
+    X: -10.0
+    Y: 10.0
+    Z: 1.0
+  NX: 10
+  NY: 10
+  NZ: 8
